Pull out separate CSS and JS files, and commonise click handler logic
